Move host-endianness magic to 'internal'.
