+++ internal/host_endianess_be.go
@@ -1,8 +1,8 @@
 //go:build arm64be
 // +build arm64be
 
-package binary
+package internal
 
 import "encoding/binary"
 
-var hostEndianess = binary.BigEndian
+var HostEndianess = binary.BigEndian
