+++ plugins/parsers/binary/entry_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/influxdata/telegraf/internal"
 	"github.com/stretchr/testify/require"
 )
 
@@ -19,7 +20,7 @@ func TestEntryConvertType(t *testing.T) {
 	testdata := []byte{0x01, 0x02, 0x03, 0x04}
 
 	e := &Entry{Type: "garbage"}
-	_, err := e.convertType(testdata, hostEndianess)
+	_, err := e.convertType(testdata, internal.HostEndianess)
 	require.EqualError(t, err, `cannot handle type "garbage"`)
 }
 
@@ -27,16 +28,16 @@ func TestEntryConvertTimeType(t *testing.T) {
 	testdata := []byte{0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09}
 
 	e := &Entry{Type: "unix_ns", location: time.UTC}
-	_, err := e.convertTimeType(testdata, hostEndianess)
+	_, err := e.convertTimeType(testdata, internal.HostEndianess)
 	require.EqualError(t, err, `too many bytes 9 vs 8`)
 }
 
 func TestConvertNumericType(t *testing.T) {
 	testdata := []byte{0x01, 0x02, 0x03, 0x04}
 
-	_, err := convertNumericType(testdata, "garbage", hostEndianess)
+	_, err := convertNumericType(testdata, "garbage", internal.HostEndianess)
 	require.EqualError(t, err, `cannot determine length for type "garbage"`)
 
-	_, err = convertNumericType(testdata, "uint8", hostEndianess)
+	_, err = convertNumericType(testdata, "uint8", internal.HostEndianess)
 	require.EqualError(t, err, `too many bytes 4 vs 1`)
 }
